more reorganization; seemingly working, tests passing
