+++ src/editor/actions/__tests__/app-message-actions.test.js
@@ -3,6 +3,8 @@ import thunk from "redux-thunk";
 import { updateAppMessages } from "../app-message-actions";
 import generateRandomId from "../../../shared/utils/generate-random-id";
 
+import { types as consoleActions } from "../../console/history/actions";
+
 jest.mock("../../../shared/utils/generate-random-id");
 
 const middlewares = [thunk];
@@ -24,7 +26,7 @@ describe("updateAppMessages", () => {
         content: "LOGGED_OUT",
         historyId: expectedRandomId,
         historyType: "APP_MESSAGE",
-        type: "ADD_TO_CONSOLE_HISTORY"
+        type: consoleActions.ADD
       },
       {
         message: {
