+++ src/editor/actions/__tests__/eval-frame-action-validator.test.js
@@ -2,6 +2,8 @@ import validateActionFromEvalFrame, {
   ActionSchemaValidationError
 } from "../eval-frame-action-validator";
 
+import { types as consoleActions } from "../../console/history/actions";
+
 describe("validateActionFromEvalFrame should throw errors as expected", () => {
   it("throw if action obj has no type property", () => {
     expect(() => validateActionFromEvalFrame({ notType: 1 })).toThrowError(
@@ -18,7 +20,7 @@ describe("validateActionFromEvalFrame should throw errors as expected", () => {
   it("throw if action obj is not valid (extra action props)", () => {
     expect(() =>
       validateActionFromEvalFrame({
-        type: "RESET_HISTORY_CURSOR",
+        type: consoleActions.UPDATE,
         extra_prop: "extra_prop is not valid"
       })
     ).toThrowError(ActionSchemaValidationError);
@@ -43,18 +45,10 @@ describe("validateActionFromEvalFrame should throw errors as expected", () => {
 });
 
 describe("validateActionFromEvalFrame should return true it action is valid", () => {
-  it("no action params", () => {
-    expect(
-      validateActionFromEvalFrame({
-        type: "RESET_HISTORY_CURSOR"
-      })
-    ).toEqual(true);
-  });
-
   it("some action params", () => {
     expect(
       validateActionFromEvalFrame({
-        type: "UPDATE_VALUE_IN_HISTORY",
+        type: consoleActions.UPDATE,
         historyItem: {
           historyId: "s03nv9dns",
           content: "ok",
@@ -64,7 +58,7 @@ describe("validateActionFromEvalFrame should return true it action is valid", ()
     ).toEqual(true);
     expect(
       validateActionFromEvalFrame({
-        type: "ADD_TO_CONSOLE_HISTORY",
+        type: consoleActions.ADD,
         historyId: "s03nv9dns",
         content: "ok",
         historyType: "CONSOLE_MESSAGE",
