+++ src/editor/console/base-icon.jsx
@@ -1,5 +1,5 @@
 import styled from "react-emotion";
-
+// fixme should this be in "shared" or should the name be specified as ConsoleIcon or something?
 const BaseIcon = (icon, style = {}) => styled(icon)`
   width: 14px;
   height: 14px;
