+++ src/editor/console/empty-pane-contents.jsx
