+++ src/editor/console/history/actions.js
@@ -1,9 +1,15 @@
-import generateRandomId from "../../shared/utils/generate-random-id";
+import generateRandomId from "../../../shared/utils/generate-random-id";
+
+export const types = {
+  ADD: "console/history/ADD",
+  UPDATE: "console/history/UPDATE",
+  UPDATE_LINE: "console/history/UPDATE_LINE"
+};
 
 export const addToConsoleHistory = (
   historyItem,
   historyId = generateRandomId()
-) => ({ type: "ADD_TO_CONSOLE_HISTORY", historyId, ...historyItem });
+) => ({ type: types.ADD, historyId, ...historyItem });
 
 export const addAppMessageToConsoleHistory = content =>
   addToConsoleHistory({ content, historyType: "APP_MESSAGE" });
@@ -35,3 +41,19 @@ export const addPluginParseErrorToHistory = errorMessage =>
     content: `plugin definition failed to parse:\n${errorMessage}`,
     level: "ERROR"
   });
+
+export const updateHistoryEntryLevel = (historyId, level) => ({
+  type: types.UPDATE,
+  historyItem: { historyId, level }
+});
+
+export const updateHistoryLineContent = (
+  historyId,
+  lineIndex,
+  lineContent
+) => ({
+  type: types.UPDATE_LINE,
+  historyId,
+  lineIndex,
+  lineContent
+});
