+++ src/eval-frame/actions/editor-message-senders.js
@@ -28,7 +28,7 @@ export function addConsoleEntryInEditor(historyItemAction) {
   const historyId = generateRandomId();
   messagePasserEval.postMessage("REDUX_ACTION", {
     historyId,
-    type: "ADD_TO_CONSOLE_HISTORY",
+    type: "console/history/ADD",
     ...historyItemAction
   });
   return historyId;
@@ -36,7 +36,7 @@ export function addConsoleEntryInEditor(historyItemAction) {
 
 export function updateConsoleEntryInEditor(historyItem) {
   messagePasserEval.postMessage("REDUX_ACTION", {
-    type: "UPDATE_VALUE_IN_HISTORY",
+    type: "console/history/UPDATE",
     historyItem
   });
 }
